backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.5
Pillow==9.5.0
drf-yasg==1.21.5
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
urllib3==2.2.1
requests==2.31.0 